[address_range]: fixed bugs in subnet and host IP generation.
Details: subnet address was being generated incorrectly due to the order of bytes messup.
Also, the mask we apply to get the host address was not applied correctly either.
Thus we were generating a range of incorrect IP addresses and then pinging them (lol).

Additional fixes and cleanups are done:
	* got rid of address_string_to_octets() and address_octets_to_string()
	* IP address conversion from string to network byte order (and back) is now done with inet_aton() and inet_ntoa()